+++ Agents/resourceAgent/RobotAgent.java
@@ -94,7 +94,7 @@ public class RobotAgent implements ResourceAgent {
 
 	@Override
 	public void teamQuery(ProductAgent productAgent, PhysicalProperty desiredProperty, ProductState currentNode, 
-			int maxTime, DirectedSparseGraph<ProductState,ResourceEvent> bid, int currentTime) {
+			double maxTime, DirectedSparseGraph<ProductState,ResourceEvent> bid, int currentTime) {
 		
 		new ResourceAgentHelper().teamQuery(productAgent, desiredProperty, currentNode, maxTime, bid,
 				currentTime, this, neighbors, tableNeighborNode, robotCapabilities, weightTransformer);
